Move querying logic to POD level rather than entries-level
