+++ packages/lib/podspec/src/index.ts
@@ -1,9 +1,6 @@
 import * as p from "./core";
-import type {
-  PodspecEntriesSerializedDef,
-  RawEntriesType
-} from "./types/entries";
-export type GenericSerializedEntriesSpec =
-  PodspecEntriesSerializedDef<RawEntriesType>;
+import type { RawEntriesType } from "./types/entries";
+import { SerializedPodspecPOD } from "./types/pod";
+export type GenericSerializedPodspecPOD = SerializedPodspecPOD<RawEntriesType>;
 export { p };
 export default p;
